Restructure druid-shell modules
This adds a platform module; all platform-specific code should
be in per-platform modules underneath it. The active platform
is reexported as a glob, and so is available at `druid_shell::platform`.

The motivation for this is to clearly separate platform code from
general code; in the previous setup it was possible for this to be ambiguous.

This is motivated by #310.
