+++ druid-shell/src/platform/mod.rs
@@ -1,0 +1,30 @@
+// Copyright 2019 The xi-editor Authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+//! Platform specific implementations.
+
+#[cfg(all(target_os = "windows", not(feature = "use_gtk")))]
+mod windows;
+#[cfg(all(target_os = "windows", not(feature = "use_gtk")))]
+pub use windows::*;
+
+#[cfg(all(target_os = "macos", not(feature = "use_gtk")))]
+mod mac;
+#[cfg(all(target_os = "macos", not(feature = "use_gtk")))]
+pub use mac::*;
+
+#[cfg(any(feature = "use_gtk", target_os = "linux"))]
+mod gtk;
+#[cfg(any(feature = "use_gtk", target_os = "linux"))]
+pub use self::gtk::*;
