+++ druid-shell/src/platform/windows/paint.rs
