+++ druid-shell/src/platform/windows/timers.rs
